pyproject.toml
[project]
name = "hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain==0.3.27",
    "langchain-community>=0.3.31",
    "langchain-google-genai>=2.1.12",
    "langchain-tavily>=0.2.12",
    "python-dotenv>=1.2.1",
    "quart>=0.20.0",
    "quart-cors>=0.8.0",
    "tavily-python>=0.7.12",
]
